request params now is extension of Encodable
