+++ swiftui_project/swiftui_project/Supports&Extensions/EncodableExtensions/ExtensionEncodable.swift
@@ -1,22 +1,10 @@
 import Foundation
 
-struct MovieParameters: Codable {
-    let randomGenreID: String
-    let page: String = "\(Int.random(in: 1...100))"
-    
-    enum CodingKeys: String, CodingKey  {
-        case randomGenreID = "with_genres"
-        case page
-    }
-}
-
-struct RequestParameters<T: Codable> {
-    let parameter: T
-    
+extension Encodable {
     func convertToDictionary() -> [String: Any]? {
         do {
             let encoder = JSONEncoder()
-            let data = try encoder.encode(self.parameter)
+            let data = try encoder.encode(self)
             guard let dictionary = try
                     JSONSerialization.jsonObject(with: data,
                                                  options: .allowFragments) as? [String: Any]
@@ -30,4 +18,3 @@ struct RequestParameters<T: Codable> {
         }
     }
 }
-
