+++ swiftui_project/swiftui_project/Supports&Extensions/MovieModel/MovieParameters.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+struct MovieParameters: Codable {
+    let randomGenreID: String
+    let page: String = "\(Int.random(in: 1...100))"
+    
+    enum CodingKeys: String, CodingKey  {
+        case randomGenreID = "with_genres"
+        case page
+    }
+}
+
+
+
